pyproject.toml
[project]
name = "signal-mcp-client"
version = "0.1.0"
description = "An MCP client that uses signal for sending and receiving texts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.3.0",
    "python-dotenv>=1.0.1",
    "websockets",
    "requests",
    "litellm",
]
